Make 'struct x11_data' common (with some ifdefs) between hw_X11, hw_xft and hw_gfx.
