+++ server/hw/hw_xft.c
@@ -1,2 +1,2 @@
-#define ENABLE_HW_XFT
+#define HW_XFT
 #include "hw_X11.c"
